refactor: move Sprites.jsx inline panels into client/src/components/sprites/ (#3390)
